Teams Creation
Getting the list of teams, their names, and the link to the team page is finished.

Next, clicking the botton to change the table and go to the next list of teams to get the team names and links to the team webpage. Storage for all the players and their SR should be done soon in class objects to make traversing arrays much easier